+++ RELEASENOTES.md
@@ -1,2 +1,13 @@
+## 1.1.1
+Added missed handling of unary expressions for given property expressions.\
+More tests.\
+`ValidateAddCustom` now accepts all properties, not only of type string.
+
+## 1.1.0
+Added `ValidateAddCustom` extension to add self-validated (outside this package) validation message to collection.\
+Enforced nullability.\
+Fixed IntelliSense Docs.\
+.Net 7 for internal tests.
+
 ## 1.0.0
 Initial release. Targets .Net Standard 2.0